[bugfix][patch][IMS 294263]헬름 리포지터리 상세페이지 차트 탭 무한로딩 수정
